Fully functional

+++ src/main/kotlin/SurfaceMath.kt
@@ -3,7 +3,7 @@ import kotlin.math.pow
 object SurfaceMath {
 
     fun binomCoeff(n: Int, i: Int): Double =
-        factorial(n).toDouble() / (factorial(i) * factorial(n - i))
+        factorial(n).toDouble() / (factorial(i).toDouble() * factorial(n - i))
 
     fun bernsteinPoly(n: Int, i: Int, u: Double) =
         binomCoeff(n, i) * (u.pow(i)) * (1 - u).pow(n - i)
